+++ packages/create-typink/templates/default/ui/src/components/shared/AccountSelection.tsx.template.ejs
@@ -82,7 +82,7 @@ export default function AccountSelection() {
 }
 <% } %>
 
-<% if (options.walletConnector === 'Subconnect') { %>
+<% if (options.walletConnector === 'SubConnect V2') { %>
 import { Box, Button, Flex, Menu, MenuButton, MenuDivider, MenuItem, MenuList, Text } from '@chakra-ui/react';
 import { useEffect, useMemo } from 'react';
 import { useWalletConnector } from '@/providers/WalletConnectorProvider.tsx';
