+++ packages/create-typink/templates/default/ui/src/components/shared/WalletSelection.tsx.template.ejs
@@ -98,7 +98,7 @@ export default function WalletSelection({
 }
 <% } %>
 
-<% if (options.walletConnector === 'Subconnect') { %>
+<% if (options.walletConnector === 'SubConnect V2') { %>
 import { Button, ChakraProps, ThemingProps } from '@chakra-ui/react';
 import { useWalletConnector } from '@/providers/WalletConnectorProvider.tsx';
 import { Props } from 'typink';
