+++ packages/create-typink/templates/default/ui/src/main.tsx.template.ejs
@@ -43,7 +43,7 @@ root.render(
 );
 <% } %>
 
-<% if (options.walletConnector === 'Subconnect' || options.walletConnector === 'Talisman Connect') { %>
+<% if (options.walletConnector === 'SubConnect V2' || options.walletConnector === 'Talisman Connect') { %>
 import { WalletConnectorProvider, useWalletConnector } from '@/providers/WalletConnectorProvider.tsx';
 
 function TypinkApp() {
