+++ packages/create-typink/templates/default/ui/src/providers/WalletConnectorProvider.tsx.template.ejs
@@ -1,4 +1,4 @@
-<% if (options.walletConnector === 'Subconnect') { %>
+<% if (options.walletConnector === 'SubConnect V2') { %>
 import { useConnectWallet, Web3OnboardProvider } from '@subwallet-connect/react';
 import { init } from '@subwallet-connect/react';
 import { createContext, useContext } from 'react';
